6 issue

+++ src/bot/dialogue/contexts/inititlize_rep.py
@@ -8,12 +8,12 @@ from aiogram.types import (
     Message,
 )
 
-from bot.dialogue.contexts.base import BaseContext, CallbackName
+from bot.dialogue.contexts.base import BaseContext, BaseSubContext, CallbackName
 from bot.user_reps import initialize_user_repository
 from sec_store.key import hash_key
 
-_CREATE_REPOSITORY_CALLBACK = CallbackName("initialize_repository")
-_CANCEL_REPOSITORY_INITIALIZATION = CallbackName("cancel_repo_initialization")
+_CREATE_REPOSITORY_CALLBACK = CallbackName("_CREATE_REPOSITORY_CALLBACK")
+_CANCEL_PASSWORD_INPUT = CallbackName("_CANCEL_PASSWORD_INPUT")
 
 
 class InitializeRepCtx(BaseContext):
@@ -21,56 +21,25 @@ class InitializeRepCtx(BaseContext):
         super().__init__(bot, user_id)
 
         self._password_input: _PasswordInput | None = None
-        self._rep_creation_message: Message | None = None
 
         create_task(
             self._send_init_keyboard(),
             name=f"Send initialize rep keyboard for {user_id=}",
         )
 
-    async def _handle_sub_ctx_result(self, sub_ctx: None):
-        raise NotImplementedError
+    async def _handle_sub_ctx_result(self, sub_ctx: "_PasswordInput"):
+        if sub_ctx.result is not None:
+            initialize_user_repository(self._user_id, hash_key(sub_ctx.result))
+            await self._bot.send_message(self._user_id, "Репозиторий успешно создан!")
+
+        self._set_ctx_over()
 
     async def _handle_message(self, message: Message):
-        if self._password_input is None:
-            await message.delete()
-            return
-
-        await self._password_input.handle_message(message)
-        result_password = self._password_input.result_password
-        if result_password is None:
-            return
-
-        initialize_user_repository(self._user_id, hash_key(result_password))
-        await self._bot.send_message(self._user_id, "Репозиторий успешно создан!")
-        if self._rep_creation_message is not None:
-            await self._rep_creation_message.delete()
-
-        self._set_ctx_over()
-
-    async def _cancel_repository_creation_callback(self, callback_query: CallbackQuery):
-        self._set_ctx_over()
-        await callback_query.message.edit_text(
-            "Создание репозитория отменено", reply_markup=None
-        )
+        await message.delete()
 
     async def _create_repositiry_callback(self, callback_query: CallbackQuery):
-        keyboard_markup = InlineKeyboardMarkup(row_width=3)
-        keyboard_markup.row(
-            InlineKeyboardButton(
-                "Отменить создание репозитория",
-                callback_data=_CANCEL_REPOSITORY_INITIALIZATION,
-            )
-        )
-
-        self._rep_creation_message = await callback_query.message.edit_text(
-            "Создание репозитория",
-            reply_markup=keyboard_markup,
-        )
-        self._set_callback(
-            _CANCEL_REPOSITORY_INITIALIZATION, self._cancel_repository_creation_callback
-        )
-        self._password_input = _PasswordInput(self._bot, self._user_id)
+        self._set_sub_ctx(_PasswordInput(self._bot, self._user_id))
+        await callback_query.message.delete()
 
     async def _send_init_keyboard(self):
         keyboard_markup = InlineKeyboardMarkup(row_width=3)
@@ -90,35 +59,47 @@ class InitializeRepCtx(BaseContext):
         )
 
 
-class _PasswordInput:
-    def __init__(self, bot: Bot, user_id: str) -> None:
-        self._bot = bot
-        self._user_id = user_id
+class _PasswordInput(BaseSubContext[str | None]):
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
 
-        create_task(
-            self._send_hello_message(), name=f"Password enter message {user_id=}"
-        )
-
+        self._password_creation_message: Message | None = None
         self._enter_password_message: Message | None = None
 
         self._password_1 = None
         self._password_2 = None
 
-    @property
-    def result_password(self) -> str | None:
-        if self._password_1 == self._password_2:
-            return self._password_1
-
-        return None
+        create_task(
+            self._send_hello_message(), name=f"Password enter message {self._user_id=}"
+        )
 
     async def _send_hello_message(self):
+        self._password_creation_message = await self._bot.send_message(
+            self._user_id,
+            "Создание пароля",
+            reply_markup=InlineKeyboardMarkup().row(
+                InlineKeyboardButton(
+                    "Отменить создание пароля",
+                    callback_data=_CANCEL_PASSWORD_INPUT,
+                )
+            ),
+        )
+        self._set_callback(_CANCEL_PASSWORD_INPUT, self._cancel_password_input)
         self._enter_password_message = await self._bot.send_message(
             self._user_id, "Придумайте пароль"
         )
 
-    async def handle_message(self, message: Message):
+    async def _cancel_password_input(self, callback_query: CallbackQuery):
+        self._set_result(None)
+
+        if self._enter_password_message is not None:
+            await self._enter_password_message.delete()
+        await callback_query.message.delete()
+
+    async def _handle_message(self, message: Message):
         await message.delete()
-        await self._enter_password_message.delete()
+        if self._enter_password_message is not None:
+            await self._enter_password_message.delete()
 
         if self._password_1 is None:
             self._password_1 = message.text
@@ -132,3 +113,9 @@ class _PasswordInput:
             self._enter_password_message = await self._bot.send_message(
                 self._user_id, "Пароли не совпадают. Попробуйте еще раз"
             )
+            return
+
+        if self._password_creation_message is not None:
+            await self._password_creation_message.delete()
+
+        self._set_result(self._password_1)
